Always call ImGui.End for ImGui.Begin
The Dear ImGui documentation says that ImGui.End must ALWAYS be called for a matching ImGui.Begin, regardless of the return value. Begin and BeginChild are inconsistent with the rest of the Dear ImGui API and the appropriate End functions must ALWAYS be called.

After making this change I am no longer experiencing the crashing that I was previously. Even if this turns out to not fix the crashing, it is still a correct change to make.